requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
python-dotenv==1.0.1
reportlab==4.2.0
flask-cors==4.0.1
PyMySQL==1.1.1
cryptography==42.0.8
